Implement datepickers on the reports tab
Use bogdanp datepicker to implement datepickers for the reports tab.
In order to acheive this, it was necessary to re-structure model
initialization for both GiftCertificates and the Main module. The
initialModel function was removed, and replaced with an init function
that returns both a model and a command message. The init function
is then passed up the module hierarchy, and finally into Main.elm.
The file GiftCertificates/Init.elm was added to avoid circular
dependencies between Models and Messages. This effectively serves as
Main.elm for the GiftCertificates submodule. The Main module does not
need Init.elm as the init function is kept in Main.elm.