Fetch username instead of issuer_id on gc queries
The query to fetch todays gift certificates was modified
and joined Gift Certificates to Users in order to return
the username of the issuer, not simply the id. On the
client side, the commands to decode the gift certificate
data were modified to handle username strings instead of
issuer_id strings. The view was modified again to display
strings instead of numeric data. Finally, the datatype
for gift_certificate_id was changed back from string to
an int.